seperate feature bar and table in HasBucket
