+++ src/components/FeatureContent/index.css
@@ -1,0 +1,9 @@
+.feature-bar {
+  display: block;
+  width: 100%;
+  height: 24px;
+  margin: 0 0 10px 0;
+  padding: 0;
+  position: relative;
+  line-height: 24px;
+}
