+++ src/components/HasBucket/index.css
@@ -10,12 +10,9 @@
 }
 .feature-container {
   display: block;
-  width: 100%;
-  height: 24px;
-  margin: 0 0 10px 0;
-  padding: 0;
   position: relative;
-  line-height: 24px;
+  width: 100%;
+  height: 68px;
 }
 .table-container {
   display: block;
